+++ lib/formex/form.ex
@@ -18,9 +18,9 @@ defmodule Formex.Form do
     * `:params` - sent parameters
     * `:phoenix_form` - `%Phoenix.HTML.Form{}`
     * `:template` - the module that implements `Formex.Template`, for example:
-      `Formex.Template.BootstrapHorizontal`. Can be set via a `Formex.View.formex_form_for` options
+    `Formex.Template.BootstrapHorizontal`. Can be set via a `Formex.View.formex_form_for` options
     * `:method` - `:post`, `:put` etc. May be used by `Formex.View`.
-      E.g. `Formex.Ecto.Builder` sets here `:put` if we editing `struct`, `:post` otherwise.
+    E.g. `Formex.Ecto.Builder` sets here `:put` if we editing `struct`, `:post` otherwise.
     * `:submitted?` - is form submitted? Set by `Formex.Controller.handle_form/1`
     * `:opts` - additional data passed in a controller. See: `Formex.Builder.create_form/5`
   """
@@ -30,16 +30,16 @@ defmodule Formex.Form do
     struct_module: nil,
     struct_info: nil,
     valid?: false,
-    errors: [],
     items: [],
     params: %{},
-    mapped_params: %{},
     phoenix_form: nil,
-    method: nil,
-    opts: [],
     template: nil,
+    method: nil,
+    submitted?: false,
+    opts: [],
+    errors: [],
     template_options: nil,
-    submitted?: false
+    mapped_params: %{}
 
   @type t :: %Form{}
 
@@ -125,6 +125,7 @@ defmodule Formex.Form do
     |> Enum.filter(&(&1.__struct__ == FormCollection))
   end
 
+  @doc false
   @spec start_creating(form :: Form.t, type :: any, name :: Atom.t, opts :: Map.t) :: Form.t
   def start_creating(form, type, name, opts \\ []) do
     info = form.struct_info[name]
@@ -136,6 +137,7 @@ defmodule Formex.Form do
     end
   end
 
+  @doc false
   @spec finish_creating(form :: Form.t) :: Form.t
   def finish_creating(form) do
     new_items = form.items
@@ -170,6 +172,41 @@ defmodule Formex.Form do
   @spec is_assoc(form :: Form.t, name :: Atom.t) :: boolean
   def is_assoc(form, name) do
     form.struct_module.__schema__(:association, name) != nil
+  end
+
+  # applies function for every select field
+  @doc false
+  @spec modify_selects_recursively(form :: Form.t, fun :: (Field.t, Form.t -> Field.t)) :: Form.t
+  def modify_selects_recursively(form, fun) do
+
+    form_items = Enum.map(form.items, fn item ->
+
+      case item do
+        collection = %FormCollection{} ->
+          forms = collection.forms
+          |> Enum.with_index()
+          |> Enum.map(fn {nested, index} ->
+            form = modify_selects_recursively(nested.form, fun)
+
+            %{nested | form: form}
+          end)
+          %{collection | forms: forms}
+
+        nested = %FormNested{} ->
+          %{nested | form: modify_selects_recursively(nested.form, fun)}
+
+        field = %Field{} ->
+          if field.type in [:select, :multiple_select] do
+            fun.(form, field)
+          else
+            field
+          end
+
+        _ -> item
+        end
+      end)
+
+    Map.put(form, :items, form_items)
   end
 
 end